+++ sdk/appconfiguration/arm-appconfiguration/src/operations/keyValues.ts
@@ -11,15 +11,12 @@ import * as coreClient from "@azure/core-client";
 import * as Mappers from "../models/mappers.js";
 import * as Parameters from "../models/parameters.js";
 import { AppConfigurationManagementClient } from "../appConfigurationManagementClient.js";
-import {
-  SimplePollerLike,
-  OperationState,
-  createHttpPoller,
-} from "@azure/core-lro";
+import { SimplePollerLike, OperationState, createHttpPoller } from "@azure/core-lro";
 import { createLroSpec } from "../lroImpl.js";
 import {
   KeyValuesGetOptionalParams,
   KeyValuesGetResponse,
+  KeyValue,
   KeyValuesCreateOrUpdateOptionalParams,
   KeyValuesCreateOrUpdateResponse,
   KeyValuesDeleteOptionalParams,
@@ -66,16 +63,24 @@ export class KeyValuesImpl implements KeyValues {
    * @param configStoreName The name of the configuration store.
    * @param keyValueName Identifier of key and label combination. Key and label are joined by $
    *                     character. Label is optional.
+   * @param keyValueParameters The parameters for creating a key-value.
    * @param options The options parameters.
    */
   createOrUpdate(
     resourceGroupName: string,
     configStoreName: string,
     keyValueName: string,
+    keyValueParameters: KeyValue,
     options?: KeyValuesCreateOrUpdateOptionalParams,
   ): Promise<KeyValuesCreateOrUpdateResponse> {
     return this.client.sendOperationRequest(
-      { resourceGroupName, configStoreName, keyValueName, options },
+      {
+        resourceGroupName,
+        configStoreName,
+        keyValueName,
+        keyValueParameters,
+        options,
+      },
       createOrUpdateOperationSpec,
     );
   }
@@ -105,8 +110,7 @@ export class KeyValuesImpl implements KeyValues {
       args: coreClient.OperationArguments,
       spec: coreClient.OperationSpec,
     ) => {
-      let currentRawResponse: coreClient.FullOperationResponse | undefined =
-        undefined;
+      let currentRawResponse: coreClient.FullOperationResponse | undefined = undefined;
       const providedCallback = args.options?.onResponse;
       const callback: coreClient.RawResponseCallback = (
         rawResponse: coreClient.FullOperationResponse,